Test esgf and urs authentication using mock
Mock testing the ESGF and URS authentication methods for connecting to an OPeNDAP server. Also updated inline docs to mention that the DAP_USER and DAP_PASSWORD environment variables need to be set for authentication to work. As pydap is not available in the default conda channel, I've removed it from the environment-py36-default.yml file.