opendap driver: use no authentication as default
Apart from historic reasons, using ESGF as the default authentication
method does not seem to be a straight forward choice. Changing the
default to no authentication most likely is the more user friendly
alternative. Apart from that and following #57, the default if no
authentication method is given should be anonymous access, but currently
this fall back does not work (see also #74).