Pass engine to xr.open_mfdataset in OpenDapSource
When OpenDapSource._open_dataset is called when urlpath is a list of
OPeNDAP URLs, no engine is specified in the call to xr.open_mfdataset
even if engine is set in the class. This prevents clients from using a
specified engine and reverts to xarray's default engine selection.

This commit changes the call to xr.open_mfdataset to include the engine
argument as the self.engine value.